KeepAliveDaemon and dead lock fixes
Updated the logic for the keep alive to guarantee that it stops sending
the keep alives once the owner TaskExecutionContext has been garbage
collected. So if the task dies with a a ThreadAbortException and cannot
signal the keep alive to stop, the keep alive will stop once the context
has been GCed.

Added a load test that provoked previously unseen dead locks when
getting the execution and critical section tokens. Fixed this by adding
new indexes and updating the queries to guarantee row level locking
isolation.